requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Werkzeug>=3.0.1
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.5.0
gspread>=5.12.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
email-validator>=2.0.0
pytz>=2024.1

